Fix: Advanced Editor upload failed if MySQL was in strict mode
If MySQL is in strict mode it will reject uploads from the advanced
editor with the following error

> PHP Fatal error: Incorrect integer value: '' for column 'ThumbWidth'
  at row 1|Gdn_Database|Query|insert GDN_Media

This is because the advanced editor is trying to insert a string value
into an integer. Setting a default value of 0 for the thumb and image
height and width values corrects this.

* Forum thread on this topic:
  https://vanillaforums.org/discussion/32255/advanced-editor-upload-failing-incorrect-integer-value#latest